Routes created with generic route to show basic navigation on dashboard if not logged in
