+++ src/components/forms/Login.js
@@ -1,34 +1,37 @@
 import React, { useContext, useState, forwardRef } from "react";
-import { useHistory } from "react-router-dom";
 
+import { useHistory } from "react-router-dom";
 import axios from "axios";
-import UserContext from "../../Context/UserContext";
-import RegisterForm from "./RegisterForm";
-import { Col, Layout, Row } from "antd";
+import LoginForm from "./LoginForm";
+import UserContext from "../../context/UserContext";
+
+import Row from "antd/lib/row";
+import Col from "antd/lib/col";
+import Layout from "antd/lib/layout";
+import "./forms.css";
 
 const BASE_URL = process.env.Url || "http://localhost:3001";
 
-const Register = () => {
+const LoginPage = () => {
   let history = useHistory();
-
   const { setUser } = useContext(UserContext);
+
   const [error, setError] = useState("");
 
-  const onSubmit = async ({ email, firstName, lastName, password }) => {
+  const onSubmit = async ({ email, password }) => {
     try {
-      const { data } = await axios.post(`${BASE_URL}/auth/register`, {
+      const { data } = await axios.post(`${BASE_URL}/auth/login`, {
         email,
-        firstName,
-        lastName,
         password,
       });
       const { token } = data;
       setUser({ email, token });
-      history.replace("/login");
+      history.replace("/dashboard");
     } catch (error) {
       setError(error.message);
     }
   };
+
   return (
     <div>
       <Layout>
@@ -38,10 +41,10 @@ const Register = () => {
             span={12}
             style={{ backgroundColor: "#FEA000" }}
           >
-            <div>component here</div>
+            <h1>component here</h1>
           </Col>
-          <Col className="rightComponent gutter-row " span={12}>
-            <RegisterForm onSubmit={onSubmit} error={setError} />
+          <Col className=" gutter-row " span={12}>
+            <LoginForm onSubmit={onSubmit} error={setError} />
           </Col>
         </Row>
       </Layout>
@@ -49,4 +52,4 @@ const Register = () => {
   );
 };
 
-export default Register;
+export default LoginPage;
