+++ src/pages/Dashboard.js
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
-import AppContext from "../../../Context/AppContext";
-import UserContext from "../../../Context/UserContext";
-import Navigation from "../../Navigation/Navigation";
+import AppContext from "../context/AppContext";
+import UserContext from "../context/UserContext";
+import Navigation from "../components/navigation/Navigation";
 
 const Dashboard = () => {
   const { user } = useContext(UserContext);
